requirements.txt
absl-py==0.12.0
astunparse==1.6.3
bert-for-tf2==0.14.9
cached-property==1.5.2
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
flatbuffers==1.12
gast==0.4.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==2.10
importlib-metadata==4.0.1
joblib==1.0.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
Markdown==3.3.4
nltk==3.6.2
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.2.4
params-flow==0.8.2
protobuf==3.17.0
py-params==0.10.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
sentencepiece==0.1.95
six==1.15.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-addons==0.13.0
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
tensorflow-hub==0.12.0
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.60.0
typeguard==2.12.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==2.0.0
wrapt==1.12.1
zipp==3.4.1
